+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2026-02-16
+
+### Added
+- Smart auto-resume: waits for non-Spotify sound before arming auto-resume.
+- New tray menu options for "Wait for Sound" and "Sound Duration".
+- Codebase formatted with black and isort.
+
 ## [0.2.0] - 2026-02-16
 
 ### Added